Add subprocess module and implement start_node method
This commit adds the subprocess module and implements the `start_node` method in the `BootStrapper` class. The `start_node` method starts a new Python process using the `subprocess.Popen` function and waits for 5 seconds before continuing. This method is used to start the AuthNodes and ContentNodes in the `service_node_checker` method.